+++ src/Response.php
@@ -5,6 +5,7 @@ namespace Spawnia\Sailor;
 use GraphQL\Executor\ExecutionResult;
 use Psr\Http\Message\ResponseInterface;
 use Safe\Exceptions\JsonException;
+use stdClass;
 
 /**
  * Represents a response sent by a GraphQL server.
@@ -18,19 +19,19 @@ class Response
     /**
      * The result of the execution of the requested operation.
      */
-    public ?\stdClass $data;
+    public ?stdClass $data;
 
     /**
-     * A non‐empty list of errors, where each error is a map.
+     * Non‐empty list of errors, where each error is a map.
      *
-     * @var array<int, \stdClass>|null
+     * @var array<int, stdClass>|null
      */
     public ?array $errors;
 
     /**
      * This entry, if set, must have a map as its value.
      */
-    public ?\stdClass $extensions;
+    public ?stdClass $extensions;
 
     public static function fromResponseInterface(ResponseInterface $response): self
     {
@@ -58,14 +59,14 @@ class Response
             throw new InvalidDataException("Received a response that is invalid JSON: {$json}", 0, $jsonException);
         }
 
-        if (! $response instanceof \stdClass) {
+        if (! $response instanceof stdClass) {
             throw new InvalidDataException("A response to a GraphQL operation must be a map, got: {$json}");
         }
 
         return self::fromStdClass($response);
     }
 
-    public static function fromStdClass(\stdClass $rawResponse): self
+    public static function fromStdClass(stdClass $rawResponse): self
     {
         $hasData = property_exists($rawResponse, 'data');
         $hasErrors = property_exists($rawResponse, 'errors');
@@ -134,7 +135,7 @@ class Response
         }
 
         foreach ($errors as $error) {
-            if (! $error instanceof \stdClass) {
+            if (! $error instanceof stdClass) {
                 throw new InvalidDataException('Each error in the response must be a map, got: ' . \Safe\json_encode($error));
             }
 
@@ -158,7 +159,7 @@ class Response
     protected static function validateData($data): void
     {
         if (
-            $data instanceof \stdClass
+            $data instanceof stdClass
             || null === $data
         ) {
             return;
@@ -176,7 +177,7 @@ class Response
      */
     protected static function validateExtensions($extensions): void
     {
-        if (! $extensions instanceof \stdClass) {
+        if (! $extensions instanceof stdClass) {
             throw new InvalidDataException('The response entry "extensions" must be a map.');
         }
     }